+++ Magento_Downloadable/requirejs-config.js
@@ -1,0 +1,9 @@
+var config = {
+  config: {
+    mixins: {
+      "Magento_Downloadable/js/downloadable": {
+        "Magento_Downloadable/js/downloadable-widget-mixin": true
+      }
+    }
+  }
+};